# src/lib/server/openapi.product.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Maildrill Product API",
    "version": "0.1.0",
    "description": "Subscribers, lists, segments, tags, templates, suppressions, auth, and campaign send. Auth: x-api-key or Bearer JWT."
  },
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" },
      "bearer": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/v1/subscribers": {
      "post": {
        "summary": "Create or update a subscriber (upsert by email)",
        "tags": ["Subscribers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string", "minLength": 1 },
                  "name": { "type": "string" },
                  "attributes": { "type": "object", "additionalProperties": {} },
                  "status": {
                    "type": "string",
                    "enum": ["active", "unsubscribed", "bounced", "complained"]
                  }
                },
                "required": ["email"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      },
      "get": {
        "summary": "List subscribers",
        "tags": ["Subscribers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["active", "unsubscribed", "bounced", "complained"]
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": { "type": "integer", "exclusiveMinimum": true, "minimum": 0, "maximum": 200 },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": { "type": "integer", "minimum": 0 },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/subscribers/{id}": {
      "get": {
        "summary": "Get a subscriber",
        "tags": ["Subscribers"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "patch": {
        "summary": "Update a subscriber",
        "tags": ["Subscribers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "nullable": true },
                  "phone": { "type": "string", "nullable": true },
                  "status": {
                    "type": "string",
                    "enum": ["active", "unsubscribed", "bounced", "complained"]
                  },
                  "attributes": { "type": "object", "additionalProperties": {} }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete a subscriber",
        "tags": ["Subscribers"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/subscribers/{id}/unsubscribe": {
      "post": {
        "summary": "Unsubscribe a subscriber",
        "tags": ["Subscribers"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/subscribers/{id}/tags/{tagId}": {
      "post": {
        "summary": "Assign a tag to a subscriber",
        "tags": ["Subscribers"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "tagId",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Remove a tag from a subscriber",
        "tags": ["Subscribers"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "tagId",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
    "/v1/templates": {
      "post": {
        "summary": "Create a template",
        "tags": ["Templates"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "channel": { "type": "string", "enum": ["email", "sms", "whatsapp", "voice"] },
                  "subject": { "type": "string", "nullable": true },
                  "preheader": { "type": "string", "nullable": true },
                  "html": { "type": "string", "nullable": true },
                  "text": { "type": "string", "nullable": true },
                  "builderDoc": { "type": "object", "additionalProperties": {}, "nullable": true },
                  "category": { "type": "string", "nullable": true },
                  "favorite": { "type": "boolean" }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": { "200": { "description": "Default Response" } }
      },
      "get": {
        "summary": "List templates",
        "tags": ["Templates"],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/v1/templates/{id}": {
      "get": {
        "summary": "Get a template",
        "tags": ["Templates"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "patch": {
        "summary": "Update a template",
        "tags": ["Templates"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 1 },
                  "channel": { "type": "string", "enum": ["email", "sms", "whatsapp", "voice"] },
                  "subject": { "type": "string", "nullable": true },
                  "preheader": { "type": "string", "nullable": true },
                  "html": { "type": "string", "nullable": true },
                  "text": { "type": "string", "nullable": true },
                  "builderDoc": { "type": "object", "additionalProperties": {}, "nullable": true },
                  "category": { "type": "string", "nullable": true },
                  "favorite": { "type": "boolean" }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete a template",
        "tags": ["Templates"],
        "parameters": [
          {
            "schema": { "type": "string", "format": "uuid" },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  }
}